Buildings and Units appear on map
#Update:
Buildings and units will now display properly on the
map. These will be displayed according to their team's
color. Displayed units and buildings are labeled with text.

Started working on the game lobby and controls.

#To do:
Fix text labels so that they will resize according
to the size of the screen.