2D map can select Selectables.
#Update:
The 2D map can now select anything that implements Selectable.
The logic is in place to give a Unit the order to attack/move
to a location or for a building to change its rally point to
the given location.

#New Files:
Added a Game Lobby and a new Menu for choosing what Unit to produce.